Fix helpers test

+++ packages/database/src/migrations/20240319010415-RemoveDisasterDashboards-modifies-data.js
@@ -22,6 +22,7 @@ const removeDisasterFromSharedDashboard = async db => {
     FROM dashboard_relation
     WHERE project_codes @> ARRAY['disaster'] AND array_length(project_codes, 1) > 1;
     `);
+  if (disasterDashboardRelationsWithMultipleProjectCodes.length === 0) return;
   await Promise.all(
     disasterDashboardRelationsWithMultipleProjectCodes.map(async relation => {
       const updatedProjectCodes = relation.project_codes.filter(code => code !== 'disaster');
@@ -54,6 +55,7 @@ const getDashboardRelations = async (db, dashboardIds) => {
 
 // Remove dashboard relations that are related to disasters
 const removeDashboardRelations = async (db, dashboardIds) => {
+  if (dashboardIds.length === 0) return;
   await db.runSql(`
     DELETE from dashboard_relation
     where dashboard_id IN (${convertArrayToString(dashboardIds)});
@@ -62,6 +64,7 @@ const removeDashboardRelations = async (db, dashboardIds) => {
 
 // Remove dashboards that are related to disasters
 const removeDashboards = async (db, dashboardIds) => {
+  if (dashboardIds.length === 0) return;
   await db.runSql(`
     DELETE from dashboard
     where id IN (${convertArrayToString(dashboardIds)});
@@ -70,6 +73,7 @@ const removeDashboards = async (db, dashboardIds) => {
 
 // Remove dashboard items that are related to disasters and don't have any other dashboard relations outside of the disaster dashboards
 const removeDisasterDashboardItems = async (db, dashboardIds, dashboardItemIds) => {
+  if (dashboardIds.length === 0 || dashboardItemIds.length === 0) return;
   // get a list of dashboard items that don't have a dashboard relation outside of the disaster dashboards
   const { rows: dashboardItemsWithOtherRelations } = await db.runSql(`
     SELECT DISTINCT child_id from dashboard_relation 
@@ -102,13 +106,13 @@ exports.setup = function (options, seedLink) {
 exports.up = async function (db) {
   await removeDisasterFromSharedDashboard(db);
   const dashboards = await getDashboards(db);
+  if (dashboards.length === 0) return;
   const dashboardIds = dashboards.map(dashboard => dashboard.id);
   const dashboardRelations = await getDashboardRelations(db, dashboardIds);
   const dashboardItemIds = dashboardRelations.map(relation => relation.child_id);
   await removeDashboardRelations(db, dashboardIds);
   await removeDisasterDashboardItems(db, dashboardIds, dashboardItemIds);
   await removeDashboards(db, dashboardIds);
-  return null;
 };
 
 exports.down = function (db) {
